App/requirements.txt
gradio
torch
torchvision
numpy
pillow